Storing and retrieving pages using a unique page identifier.
ADD: Store unique page identifier in database.
ADD: Methods allowing client code to work with page identifiers.

A unique page identifier is added to a page when it is saved.
If an identifier is present in the database, that one will be used.
If an identifier is not present, it will be generated. The existing
signal-wiring will yield in the new identifier being stored in the
database.