In InsertLinkDialog translate file to page link
If the user selects a file which is a page source, show a page
link instead of of a file link. This is inended to reduce confusion
with users who use the "browse" button in the InsertLinkDialog and
select a file with the intention to link a page.

Part of #2468